[2MGFX] Move Parser to new namespace and folder. (#6110)
* Move parser to namespace TwoMGFX.TPGParser

* Move WhitespaceNodes from ParseTreeTools to ShaderResult.

ParseTreeTools is now used exlusivly by ParseTree.
WhitespaceNodes is used only by ShaderResult.

* Move Parser to TPGParser folder
