+++ Tools/2MGFX/ShaderData.mojo.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Linq;
+using TwoMGFX.TPGParser;
 
 namespace TwoMGFX
 {
