+++ Tools/2MGFX/ShaderData.pssl.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using TwoMGFX.TPGParser;
 
 namespace TwoMGFX
 {
