+++ Tools/2MGFX/ShaderData.sharpdx.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using SharpDX.Direct3D;
+using TwoMGFX.TPGParser;
 
 namespace TwoMGFX
 {
