+++ Tools/2MGFX/ShaderProfile.DirectX.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using TwoMGFX.TPGParser;
 
 namespace TwoMGFX
 {
