+++ Tools/2MGFX/ShaderProfile.cs
@@ -9,6 +9,7 @@ using System.Globalization;
 using System.Linq;
 using System.Text.RegularExpressions;
 using Microsoft.Xna.Framework.Content.Pipeline;
+using TwoMGFX.TPGParser;
 
 namespace TwoMGFX
 {
