+++ Tools/2MGFX/TPGParser/MGFX.tpg
@@ -4,7 +4,7 @@
 // https://github.com/SickheadGames/TinyPG
 //
 
-<% @TinyPG Namespace="TwoMGFX" Language="C#"%>
+<% @TinyPG Namespace="TwoMGFX.TPGParser" Language="C#"%>
 
 [Skip] BlockComment -> @"/\*([^*]|\*[^/])*\*/";
 [Skip] Comment -> @"//[^\n\r]*";
