+++ Tools/2MGFX/TPGParser/ParseTreeTools.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using Microsoft.Xna.Framework;
 
-namespace TwoMGFX
+namespace TwoMGFX.TPGParser
 {
 	public static class ParseTreeTools
 	{
@@ -57,38 +57,6 @@ namespace TwoMGFX
 	        }
 
             return new Color(r, g, b, a);
-	    }
-
-        public static void WhitespaceNodes(TokenType type, List<ParseNode> nodes, ref string sourceFile)
-        {
-            for (var i = 0; i < nodes.Count; i++)
-            {
-                var n = nodes[i];
-                if (n.Token.Type != type)
-                {
-                    WhitespaceNodes(type, n.Nodes, ref sourceFile);
-                    continue;
-                }
-
-                // Get the full content of this node.
-                var start = n.Token.StartPos;
-                var end = n.Token.EndPos;
-                var length = end - n.Token.StartPos;
-                var content = sourceFile.Substring(start, length);
-
-                // Replace the content of this node with whitespace.
-                for (var c = 0; c < length; c++)
-                {
-                    if (!char.IsWhiteSpace(content[c]))
-                        content = content.Replace(content[c], ' ');
-                }
-
-                // Add the whitespace back to the source file.
-                var newfile = sourceFile.Substring(0, start);
-                newfile += content;
-                newfile += sourceFile.Substring(end);
-                sourceFile = newfile;
-            }
-        }
+	    }        
     }
 }
