+++ Tools/2MGFX/TPGParser/TechniqueInfo.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TwoMGFX
+namespace TwoMGFX.TPGParser
 {
     public class TechniqueInfo
     {
